backend/requirements.txt
Flask==2.3.3
gunicorn==20.1.0
psycopg2-binary==2.9.7
# Add other Python dependencies here
